rdpq: add ZBUF_MAX and ZBUF_VAL macros
ZBUF_MAX is the correct clear value to use with rdpq_clear_z. It
wasn't exposed before so people had to know it, and possibly put the
magic number in the source code, that looked weird.

ZBUF_VAL allows to clear the Z-buffer with an arbitrary value, by
performing a conversion from a floating value to the low-level packed
Z format used by the RDP in the Z-buffer.