rdpq: add TEX0_BUG combiner input
Referring to the first texture (TEX0) in the second stage of a combiner
is buggy on hardware, because the RDP accesses the texture's "next
pixel" in the scanline instead.

We used to forbid creating a combiner like that altogether, but maybe
it is a bit too rough. Sometimes the 1-pixel sampling shift isn't too
bad and can still help in some cases. So, as a workaround, we now
allow for it by calling it TEX0_BUG, so that users will be aware that
there is some issue on hardware (especially given that Ares doesn't
emulate this correctly).