tools: improve binout library to add support to placeholders
binout is used by tools to generate binary files. This change adds
support for "placeholders" which is a way to leave an empty space
in a binary file to be filled up later. A very common case is storing
offsets to further sections in the file.

Moreover, the basic wNN() functions are improved to assert in case
the variable being passed does not fit the specified bitsize, to avoid
silent truncations.