wav64: make wav64_close also stop playback
Calling wav64_close on an in-progress playback would cause a crash. This
is quite unfortunate and unexpected by the user.

We add a private API to the mixer to do this, but it's pretty clear
that wav64 will have to track whether playback is in progress or not
at some point. For now, let's fix the crash.